scripts/src/requirements_era5.txt
nr_objstore_util==0.11.0
ecmwf-datastores-client==0.4.1